requirements.txt
requests 
python-dotenv 
google-adk 
flask 
google-generativeai 
langchain 
langchain-google-genai 
langchain-core